refactor: Rename admin to user to avoid conflicts
